requirements.txt
asgiref==3.2.3
Django==3.0
django-html-field==0.1.3
djangorestframework==3.10.3
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
